ref: breakup table into multiple components
