+++ src/components/Table/styles.ts
@@ -88,10 +88,6 @@ export const Wrap = styled.div<{ width: string; rows: number }>`
   }
 `
 
-export const HeaderRow = styled(Row)`
-  --row-color: #fff;
-`
-
 export const Select = styled(Cell)`
   padding: 0;
   display: flex;
@@ -101,25 +97,6 @@ export const Select = styled(Cell)`
   input {
     display: block;
     margin: 0;
-  }
-`
-
-export const Header = styled(Cell)`
-  font-weight: bold;
-  text-transform: capitalize;
-
-  &[data-sortdir='ASC'] {
-    cursor: s-resize;
-  }
-
-  &[data-sortdir='DESC'] {
-    cursor: n-resize;
-  }
-
-  svg {
-    margin-left: auto;
-    transform: scale(0.9);
-    fill: var(--cl-action-dark);
   }
 `
 
@@ -149,47 +126,6 @@ export const Item = styled.span`
 
   &[data-column='role'] {
     text-transform: lowercase;
-  }
-`
-
-export const LoadingPlaceholder = styled.div<{ rows: number }>`
-  & > div {
-    display: block;
-    width: 100%;
-    height: var(--row-height);
-    opacity: 0;
-    animation: fade 2s linear 0s infinite;
-
-    &:nth-of-type(2n) {
-      background-color: #dceffd;
-    }
-
-    &:nth-of-type(2n + 1) {
-      background-color: #cfe8fc;
-    }
-
-    /* stylelint-disable-next-line */
-    ${({ rows }) =>
-      Array(rows)
-        .fill(0)
-        .map(
-          (_, i) => `
-              &:nth-of-type(${i + 1}n) {
-                animation-delay: ${(i / (rows * 3)) * 2}s;
-              }`
-        )
-        .join('\n')}
-  }
-
-  @keyframes fade {
-    0%,
-    20% {
-      opacity: 0;
-    }
-
-    10% {
-      opacity: 0.8;
-    }
   }
 `
 
@@ -238,161 +174,5 @@ export const NavItem = styled.div`
     * ~ & {
       padding-left: 0.5rem;
     }
-  }
-`
-
-export const Dropdown = styled.div`
-  display: none;
-  position: absolute;
-  left: 0;
-  top: 100%;
-  box-shadow: 0 0 2px 1px #0005;
-  background: #fff;
-  z-index: 10;
-
-  *[aria-expanded='true'] > & {
-    display: block;
-  }
-`
-
-export const Customize = styled.div`
-  padding: 1rem;
-
-  * {
-    white-space: nowrap;
-  }
-
-  h4 {
-    margin-top: 0;
-  }
-
-  ul {
-    padding: 0;
-    margin: 0;
-  }
-
-  li {
-    list-style: none;
-    display: flex;
-    align-items: center;
-
-    input {
-      margin: 0;
-    }
-
-    label {
-      margin-left: 0.5rem;
-    }
-  }
-`
-
-export const ViewToggle = styled.div`
-  display: contents;
-
-  svg {
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    transform: scale(1) translateX(-50%) translateY(-50%);
-  }
-`
-
-export const SearchWrap = styled.div`
-  display: block;
-  position: relative;
-  margin-left: 1rem;
-
-  --collapsed-size: 15em;
-
-  flex: 1 0 var(--collapsed-size);
-`
-
-export const SearchBar = styled.form`
-  display: flex;
-  align-items: center;
-  height: 2rem;
-  border: 1px solid var(--border-color-strong);
-  border-radius: 0.25em;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  padding-left: 0.5em;
-  box-sizing: border-box;
-  width: 100%;
-
-  --transition-time: 0.2s;
-
-  transition: border-color var(--transition-time) ease;
-
-  svg {
-    width: 1em;
-    height: 1em;
-    margin-right: 0.3em;
-    fill: var(--border-color-strong);
-    transition: fill var(--transition-time) ease;
-    cursor: pointer;
-  }
-
-  &::after {
-    content: '';
-    background: #fff;
-    position: absolute;
-    right: -2px;
-    top: -2px;
-    height: calc(100% + 4px);
-    width: calc(100% - var(--collapsed-size) + 0.5em);
-    transform-origin: right;
-    transition: transform var(--transition-time) ease, opacity 0s 0.05s;
-  }
-
-  &::before {
-    content: '';
-    position: absolute;
-    right: calc(100% - var(--collapsed-size) - 1px);
-    top: -1px;
-    height: 100%;
-    width: 1rem;
-    z-index: 2;
-    border: 1px solid;
-    border-color: inherit;
-    border-left: none;
-    border-top-right-radius: inherit;
-    border-bottom-right-radius: inherit;
-    transition: right var(--transition-time) ease, opacity 0s 0s;
-  }
-
-  &[data-focus='true']::after {
-    transform: scaleX(0);
-    opacity: 0;
-    transition: transform var(--transition-time) ease,
-      opacity 0s var(--transition-time);
-  }
-
-  &[data-focus='true']::before {
-    right: 0;
-    opacity: 0;
-    transition: right var(--transition-time) ease,
-      opacity 0s var(--transition-time);
-  }
-
-  &[data-focus='true'] {
-    border-color: var(--cl-action-dark);
-  }
-
-  &[data-focus='true'] svg {
-    fill: var(--cl-action-dark);
-  }
-`
-
-export const SearchInput = styled.input`
-  flex-grow: 1;
-  border: none;
-  border-radius: inherit;
-  font-family: inherit;
-  font-size: 1em;
-  color: var(--cl-text-strong);
-
-  &:focus {
-    outline: none;
   }
 `
